requirements.txt
# Python dependencies for OpenCanvas WAT Framework

# Core utilities
python-dotenv>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Supabase
supabase>=2.0.0

# HTTP client
requests>=2.31.0

# Authentication
PyJWT>=2.8.0

# Production (optional)
gunicorn>=21.2.0
